Add tracing to run-tests.sh
